LayoutTests/js/dom/resizable-array-buffer-view-serialization-expected.txt

Resizable ArrayBuffers should be serializable

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS arrayBuffer.resizable is true
PASS arrayBuffer.byteLength is 36
PASS arrayBuffer.maxByteLength is 128
PASS int32AutoArray.length is 8
PASS int32AutoArray.byteOffset is 4
PASS cloned.buffer.resizable is true
PASS cloned.buffer.byteLength is 36
PASS cloned.buffer.maxByteLength is 128
PASS cloned.length is 8
PASS cloned.byteOffset is 4
PASS cloned.buffer.byteLength is 128
PASS cloned.buffer.maxByteLength is 128
PASS cloned.length is 31
PASS successfullyParsed is true

TEST COMPLETE
